requirements.txt
# Core dependencies
streamlit>=1.28.0
openai>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Data visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Additional utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
